refactor: Midday enum 타입 분리
