+++ frontend/src/types/time.ts
@@ -1,0 +1,9 @@
+export enum Midday {
+  AM = '오전',
+  PM = '오후',
+}
+export interface Time {
+  midday: Midday;
+  hour: number;
+  minute: number;
+}
